backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
httpx
python-dotenv
